+++ consensus/debugging.go
@@ -35,7 +35,7 @@ func (m *consensusModule) GetNodeState() typesCons.ConsensusNodeState {
 		Height:   m.height,
 		Round:    uint8(m.round),
 		Step:     uint8(m.step),
-		IsLeader: m.isLeader(),
+		IsLeader: m.IsLeader(),
 		LeaderId: leaderId,
 	}
 }
@@ -44,7 +44,7 @@ func (m *consensusModule) resetToGenesis(_ *messaging.DebugMessage) {
 	m.logger.Debug().Msg(typesCons.DebugResetToGenesis)
 
 	m.height = 0
-	m.resetForNewHeight()
+	m.ResetForNewHeight()
 	m.clearLeader()
 	m.clearMessagesPool()
 	m.GetBus().GetPersistenceModule().HandleDebugMessage(&messaging.DebugMessage{
@@ -88,32 +88,4 @@ func (m *consensusModule) togglePacemakerManualMode(_ *messaging.DebugMessage) {
 		m.logger.Debug().Str("pacemaker_mode", "AUTOMATIC").Msg("Toggle pacemaker to AUTOMATIC mode")
 	}
 	m.paceMaker.SetManualMode(newMode)
-}
-
-// This Pacemaker interface is only used for development & debugging purposes.
-type PacemakerDebug interface {
-	SetManualMode(bool)
-	IsManualMode() bool
-	ForceNextView()
-}
-
-type paceMakerDebug struct {
-	manualMode                bool
-	debugTimeBetweenStepsMsec uint64
-
-	// IMPROVE: Consider renaming to `previousRoundQC`
-	quorumCertificate *typesCons.QuorumCertificate
-}
-
-func (p *paceMaker) IsManualMode() bool {
-	return p.manualMode
-}
-
-func (p *paceMaker) SetManualMode(manualMode bool) {
-	p.manualMode = manualMode
-}
-
-func (p *paceMaker) ForceNextView() {
-	lastQC := p.quorumCertificate
-	p.startNextView(lastQC, true)
 }