+++ consensus/helpers.go
@@ -152,13 +152,6 @@ func (m *consensusModule) isOptimisticThresholdMet(numSignatures int, validators
 	return nil
 }
 
-func (m *consensusModule) resetForNewHeight() {
-	m.round = 0
-	m.block = nil
-	m.highPrepareQC = nil
-	m.lockedQC = nil
-}
-
 func protoHash(m proto.Message) string {
 	b, err := codec.GetCodec().Marshal(m)
 	if err != nil {
@@ -243,17 +236,8 @@ func (m *consensusModule) clearMessagesPool() {
 }
 
 /*** Leader Election Helpers ***/
-
-func (m *consensusModule) isLeaderUnknown() bool {
-	return m.leaderId == nil
-}
-
-func (m *consensusModule) isLeader() bool {
-	return m.leaderId != nil && *m.leaderId == m.nodeId
-}
-
 func (m *consensusModule) isReplica() bool {
-	return !m.isLeader()
+	return !m.IsLeader()
 }
 
 func (m *consensusModule) clearLeader() {
@@ -285,7 +269,7 @@ func (m *consensusModule) electNextLeader(message *typesCons.HotstuffMessage) er
 
 	idToValAddrMap := typesCons.NewActorMapper(validators).GetIdToValAddrMap()
 
-	if m.isLeader() {
+	if m.IsLeader() {
 		m.setLogPrefix("LEADER")
 		m.logger.Info().Fields(
 			map[string]any{
@@ -311,10 +295,15 @@ func (m *consensusModule) electNextLeader(message *typesCons.HotstuffMessage) er
 /*** General Infrastructure Helpers ***/
 
 func (m *consensusModule) setLogPrefix(logPrefix string) {
+<<<<<<< HEAD
 	logger.Global.UpdateFields(map[string]any{
 		"kind": logPrefix,
 	})
 	m.logger = logger.Global.CreateLoggerForModule("consensus")
+=======
+	m.logPrefix = logPrefix
+	m.paceMaker.SetLogPrefix(logPrefix)
+>>>>>>> 1d135866afd1e58e8e548b0ced55d829366112ed
 }
 
 func (m *consensusModule) getValidatorsAtHeight(height uint64) ([]*coreTypes.Actor, error) {