+++ persistence/types/protocol_account.go
@@ -1,0 +1,33 @@
+package types
+
+// Interface common to protocol Accounts and Pools at the persistence schema layer.
+// This exposes SQL specific attributes and queries.
+type ProtocolAccountSchema interface {
+	/*** Account Attributes ***/
+
+	// SQL Table Name
+	GetTableName() string
+	// SQL Column Name
+	GetAccountSpecificColName() string
+	// SQL Table Schema
+	GetTableSchema() string
+
+	/*** Read/Get Queries ***/
+
+	// Returns a query to get all accounts
+	GetAllQuery(height int64) string
+	// Returns a query to get the balance of an account at a specified height
+	GetAccountAmountQuery(identifier string, height int64) string // Identifier can either be address (cryptographic ID) (Account) or semantic name (Pool)
+	// Returns a query to select all accounts updated at a specified height
+	GetAccountsUpdatedAtHeightQuery(height int64) string
+
+	/*** Create/Insert Queries ***/
+
+	// Returns a query to insert an account amount at a specified height
+	InsertAccountQuery(identifier, amount string, height int64) string // Identifier can either be address (cryptographic ID) (Account) or semantic name (Pool)
+
+	/*** Debug Queries Only ***/
+
+	// Returns a query to clear all accounts
+	ClearAllAccounts() string
+}