+++ shared/CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.0.17] - 2023-01-27
+
+- Add `Param` and `Flag` protobufs for use in updating merkle tree
+
+## [0.0.0.16] - 2023-01-24
+
+- Add `ConsensusPacemaker` interface that is implemented by the consensus module
+
+## [0.0.0.15] - 2023-01-20
+
+- Remove `address []byte` argument from `InsertPool` function in `PostgresRWContext`
+
+## [0.0.0.14] - 2023-01-19
+
+- Rewrite `interface{}` to `any`
+
 ## [0.0.0.13] - 2023-01-18
 
 - Create `block.proto` which consolidates the definition of a `Block` protobuf under `shared/core/types`